refactor: split move actions from mover code
This opens up the possibility of triggering movement from other actions in
future, e.g. insert.
