+++ src/actions/mover.ts
@@ -4,26 +4,19 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import * as Constants from '../constants';
+import type {BlockSvg, IDragger, IDragStrategy} from 'blockly';
 import {
   ASTNode,
-  Connection,
-  ContextMenuRegistry,
-  ShortcutRegistry,
-  WorkspaceSvg,
   common,
+  Connection,
   registry,
   utils,
+  WorkspaceSvg,
 } from 'blockly';
-import type {BlockSvg, IDragger, IDragStrategy} from 'blockly';
-import {Navigation} from '../navigation';
-import {KeyboardDragStrategy} from '../keyboard_drag_strategy';
+import * as Constants from '../constants';
 import {Direction, getXYFromDirection} from '../drag_direction';
-
-const KeyCodes = utils.KeyCodes;
-const createSerializedKey = ShortcutRegistry.registry.createSerializedKey.bind(
-  ShortcutRegistry.registry,
-);
+import {KeyboardDragStrategy} from '../keyboard_drag_strategy';
+import {Navigation} from '../navigation';
 
 /**
  * The distance to move an item, in workspace coordinates, when
@@ -32,7 +25,7 @@ const createSerializedKey = ShortcutRegistry.registry.createSerializedKey.bind(
 const UNCONSTRAINED_MOVE_DISTANCE = 20;
 
 /**
- * Actions for moving blocks with keyboard shortcuts.
+ * Low-level code for moving blocks with keyboard shortcuts.
  */
 export class Mover {
   /**
@@ -57,145 +50,6 @@ export class Mover {
   private oldDragStrategy: IDragStrategy | null = null;
 
   constructor(protected navigation: Navigation) {}
-
-  private shortcuts: ShortcutRegistry.KeyboardShortcut[] = [
-    // Begin and end move.
-    {
-      name: 'Start move',
-      preconditionFn: (workspace) => this.canMove(workspace),
-      callback: (workspace) => this.startMove(workspace),
-      keyCodes: [KeyCodes.M],
-    },
-    {
-      name: 'Finish move',
-      preconditionFn: (workspace) => this.isMoving(workspace),
-      callback: (workspace) => this.finishMove(workspace),
-      keyCodes: [KeyCodes.ENTER],
-      allowCollision: true,
-    },
-    {
-      name: 'Abort move',
-      preconditionFn: (workspace) => this.isMoving(workspace),
-      callback: (workspace) => this.abortMove(workspace),
-      keyCodes: [KeyCodes.ESC],
-      allowCollision: true,
-    },
-
-    // Constrained moves.
-    {
-      name: 'Move left, constrained',
-      preconditionFn: (workspace) => this.isMoving(workspace),
-      callback: (workspace) => this.moveConstrained(workspace, Direction.Left),
-      keyCodes: [KeyCodes.LEFT],
-      allowCollision: true,
-    },
-    {
-      name: 'Move right unconstrained',
-      preconditionFn: (workspace) => this.isMoving(workspace),
-      callback: (workspace) => this.moveConstrained(workspace, Direction.Right),
-      keyCodes: [KeyCodes.RIGHT],
-      allowCollision: true,
-    },
-    {
-      name: 'Move up, constrained',
-      preconditionFn: (workspace) => this.isMoving(workspace),
-      callback: (workspace) => this.moveConstrained(workspace, Direction.Up),
-      keyCodes: [KeyCodes.UP],
-      allowCollision: true,
-    },
-    {
-      name: 'Move down constrained',
-      preconditionFn: (workspace) => this.isMoving(workspace),
-      callback: (workspace) => this.moveConstrained(workspace, Direction.Down),
-      keyCodes: [KeyCodes.DOWN],
-      allowCollision: true,
-    },
-
-    // Unconstrained moves.
-    {
-      name: 'Move left, unconstrained',
-      preconditionFn: (workspace) => this.isMoving(workspace),
-      callback: (workspace) =>
-        this.moveUnconstrained(workspace, Direction.Left),
-      keyCodes: [
-        createSerializedKey(KeyCodes.LEFT, [KeyCodes.ALT]),
-        createSerializedKey(KeyCodes.LEFT, [KeyCodes.CTRL]),
-      ],
-    },
-    {
-      name: 'Move right, unconstrained',
-      preconditionFn: (workspace) => this.isMoving(workspace),
-      callback: (workspace) =>
-        this.moveUnconstrained(workspace, Direction.Right),
-      keyCodes: [
-        createSerializedKey(KeyCodes.RIGHT, [KeyCodes.ALT]),
-        createSerializedKey(KeyCodes.RIGHT, [KeyCodes.CTRL]),
-      ],
-    },
-    {
-      name: 'Move up unconstrained',
-      preconditionFn: (workspace) => this.isMoving(workspace),
-      callback: (workspace) => this.moveUnconstrained(workspace, Direction.Up),
-      keyCodes: [
-        createSerializedKey(KeyCodes.UP, [KeyCodes.ALT]),
-        createSerializedKey(KeyCodes.UP, [KeyCodes.CTRL]),
-      ],
-    },
-    {
-      name: 'Move down, unconstrained',
-      preconditionFn: (workspace) => this.isMoving(workspace),
-      callback: (workspace) =>
-        this.moveUnconstrained(workspace, Direction.Down),
-      keyCodes: [
-        createSerializedKey(KeyCodes.DOWN, [KeyCodes.ALT]),
-        createSerializedKey(KeyCodes.DOWN, [KeyCodes.CTRL]),
-      ],
-    },
-  ];
-
-  menuItems: ContextMenuRegistry.RegistryItem[] = [
-    {
-      displayText: 'Move Block (M)',
-      preconditionFn: (scope) => {
-        const workspace = scope.block?.workspace as WorkspaceSvg | null;
-        if (!workspace) return 'hidden';
-        return this.canMove(workspace) ? 'enabled' : 'disabled';
-      },
-      callback: (scope) => {
-        const workspace = scope.block?.workspace as WorkspaceSvg | null;
-        if (!workspace) return false;
-        this.startMove(workspace);
-      },
-      scopeType: ContextMenuRegistry.ScopeType.BLOCK,
-      id: 'move',
-      weight: 8.5,
-    },
-  ];
-
-  /**
-   * Install the actions as both keyboard shortcuts and (where
-   * applicable) context menu items.
-   */
-  install() {
-    for (const shortcut of this.shortcuts) {
-      ShortcutRegistry.registry.register(shortcut);
-    }
-    for (const menuItem of this.menuItems) {
-      ContextMenuRegistry.registry.register(menuItem);
-    }
-  }
-
-  /**
-   * Uninstall these actions.
-   */
-  uninstall() {
-    for (const shortcut of this.shortcuts) {
-      ShortcutRegistry.registry.unregister(shortcut.name);
-    }
-    for (const menuItem of this.menuItems) {
-      ContextMenuRegistry.registry.unregister(menuItem.id);
-    }
-  }
 
   /**
    * Returns true iff we are able to begin moving the block which
